requirements.txt
appnope==0.1.2
asttokens==2.0.5
backcall==0.2.0
bcrypt==3.2.0
black==21.12b0
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
click==8.0.3
cryptography==36.0.1
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.3
executing==0.8.2
google-api-core==2.4.0
google-auth==2.4.1
google-cloud-core==2.2.2
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
idna==3.3
install==1.3.5
ipykernel==6.7.0
ipython==8.0.1
jedi==0.18.1
jupyter-client==7.1.2
jupyter-core==4.9.1
matplotlib-inline==0.1.3
mypy-extensions==0.4.3
nest-asyncio==1.5.4
paramiko==2.9.2
parso==0.8.3
pathspec==0.9.0
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.4.1
prompt-toolkit==3.0.24
protobuf==3.19.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
Pygments==2.11.2
PyNaCl==1.5.0
pysftp==0.2.9
python-dateutil==2.8.2
pyzmq==22.3.0
requests==2.27.1
rsa==4.8
six==1.16.0
stack-data==0.1.4
tomli==1.2.3
tornado==6.1
traitlets==5.1.1
typing_extensions==4.0.1
urllib3==1.26.8
wcwidth==0.2.5
